изменено:      CHANGELOG.md 	изменено:      README.md 	изменено:      intaro.retailcrm/classes/general/ApiClient_v5.php 	изменено:      intaro.retailcrm/classes/general/history/RetailCrmHistory_v5.php 	изменено:      intaro.retailcrm/classes/general/order/RetailCrmOrder_v5.php 	изменено:      intaro.retailcrm/include.php 	изменено:      intaro.retailcrm/install/export/local/components/intaro/lp.register/templates/.default/lang/ru/template.php 	изменено:      intaro.retailcrm/install/export/local/components/intaro/lp.register/templates/.default/result_modifier.php 	изменено:      intaro.retailcrm/install/export/local/components/intaro/lp.register/templates/.default/template.php 	изменено:      intaro.retailcrm/install/export/local/components/intaro/lp.score/component.php 	изменено:      intaro.retailcrm/install/export/local/components/intaro/lp.score/lang/ru/component.php 	изменено:      intaro.retailcrm/install/export/local/components/intaro/main.register/templates/.default/lang/ru/template.php 	изменено:      intaro.retailcrm/install/export/local/components/intaro/main.register/templates/.default/result_modifier.php 	изменено:      intaro.retailcrm/install/export/local/components/intaro/main.register/templates/.default/template.php 	изменено:      intaro.retailcrm/install/export/local/components/intaro/sale.basket.basket/templates/.default/lang/ru/result_modifier.php 	изменено:      intaro.retailcrm/install/export/local/components/intaro/sale.basket.basket/templates/.default/result_modifier.php 	изменено:      intaro.retailcrm/install/export/local/components/intaro/sale.basket.basket/templates/.default/template.php 	изменено:      intaro.retailcrm/install/export/local/components/intaro/sale.order.ajax/templates/.default/lang/ru/template.php 	изменено:      intaro.retailcrm/install/export/local/components/intaro/sale.order.ajax/templates/.default/order_ajax.js 	изменено:      intaro.retailcrm/install/export/local/components/intaro/sale.order.ajax/templates/.default/result_modifier.php 	изменено:      intaro.retailcrm/install/export/local/components/intaro/sale.order.ajax/templates/.default/scripts/intaro.js 	изменено:      intaro.retailcrm/install/export/local/components/intaro/sale.order.ajax/templates/.default/template.php 	изменено:      intaro.retailcrm/install/index.php 	изменено:      intaro.retailcrm/install/version.php 	изменено:      intaro.retailcrm/lib/component/builder/bitrix/loyaltydatabuilder.php 	изменено:      intaro.retailcrm/lib/component/configprovider.php 	изменено:      intaro.retailcrm/lib/component/constants.php 	изменено:      intaro.retailcrm/lib/component/handlers/eventshandlers.php 	изменено:      intaro.retailcrm/lib/icml/icmldirector.php 	изменено:      intaro.retailcrm/lib/icml/xmlofferbuilder.php 	изменено:      intaro.retailcrm/lib/model/bitrix/xml/xmloffer.php 	изменено:      intaro.retailcrm/lib/service/customerservice.php 	изменено:      intaro.retailcrm/lib/service/loyaltyaccountservice.php 	изменено:      intaro.retailcrm/lib/service/loyaltyservice.php 	изменено:      intaro.retailcrm/lib/service/orderloyaltydataservice.php 	изменено:      intaro.retailcrm/options.php 	изменено:      tests/BitrixTestCase.php 	изменено:      tests/classes/general/AddressBuilderTest.php 	изменено:      tests/classes/general/CustomerBuilderTest.php 	изменено:      tests/classes/general/order/RetailCrmOrder_v5Test.php 	изменено:      tests/classes/general/services/RetailCrmServiceTest.php

+++ intaro.retailcrm/lib/service/orderloyaltydataservice.php
@@ -523,7 +523,7 @@ class OrderLoyaltyDataService
             'UF_BONUS_COUNT'  => [
                 'ENTITY_ID'         => $ufObject,
                 'FIELD_NAME'        => 'UF_BONUS_COUNT',
-                'USER_TYPE_ID'      => 'integer',
+                'USER_TYPE_ID'      => 'double',
                 'MANDATORY'         => 'N',
                 'EDIT_FORM_LABEL'   => [
                     'ru' => Loc::GetMessage('UF_BONUS_COUNT', null, 'ru'),
